separating actions by entity. renamed to user
